YAY-114 refactor: change string on enum in Post, transfer hook, constans in settings
