+++ src/features/post/api/index.ts
@@ -1,3 +1,2 @@
 export * from './post.service'
 export * from './post.types'
-export * from './postsEndpoints'
