+++ src/features/settings/constants/accountType.ts
