+++ src/features/settings/constants/paymentResult.ts
