+++ src/features/settings/index.ts
@@ -1,14 +1,14 @@
-export { ACCOUNT_TYPE } from './model/constants/accountType'
-export { PAYMENT_RESULT } from './model/constants/paymentResult'
+export { ACCOUNT_TYPE } from './constants/accountType'
+export { PAYMENT_RESULT } from './constants/paymentResult'
 export { SubscriptionTabPage } from './ui/SubscriptionTabPage/SubscriptionTabPage'
 export { SubscriptionOptions } from './ui/SubscriptionTabPage/SubscriptionOptions/SubscriptionOptions'
 export { CurrentSubscription } from './ui/SubscriptionTabPage/CurrentSubscription/CurrentSubscription'
 export { DevicesTabPage } from './ui/DevicesTabPage/DevicesTabPage'
 export { InfoTabPage } from './ui/InfoTabPage/InfoTabPage'
 export { UserAccountType } from './ui/SubscriptionTabPage/UserAccountType/UserAccountType'
-export { SETTINGS_TAB_VALUES, type SettingsTabType } from './model/constants/settingTabValues'
+export { SETTINGS_TAB_VALUES, type SettingsTabType } from './constants/settingTabValues'
 export {
    useSubscription,
    type SubscriptionPlanName,
    type AccountType,
-} from './model/hooks/useSubscription'
+} from './hooks/useSubscription'
