+++ src/shared/enums/index.ts
@@ -2,3 +2,4 @@ export * from './authEndpoints'
 export * from './filesEndpoints'
 export * from './userEndpoints'
 export * from './publicPostsEndpoints'
+export * from './postsEndpoints'
