+++ src/shared/enums/postsEndpoints.ts
