+++ src/shared/enums/publicPostsEndpoints.ts
@@ -1,4 +1,5 @@
 export enum PublicPostsEndpoints {
    lastPosts = '/public/posts/last-posts',
    publicPosts = '/public/posts/users',
+   publicPost = '/public/posts',
 }
